Seperate edit product component.
