+++ src/store/actions/productActions.js
@@ -17,19 +17,16 @@ import {
 export const fetchProducts = () => async (dispatch) => {
   dispatch({ type: PRODUCT_FETCHING });
   const { data } = await axios.get("/api/products");
-  // console.log(data);
   dispatch({ type: PRODUCT_FETCHED, payload: data.products });
 };
 
 export const addProduct = (productData) => async (dispatch) => {
   dispatch({ type: CREATE_PRODUCT_REQUEST, payload: productData });
-  console.log(productData);
   const { data } = await axios.post("/api/products/create", productData, {
     headers: {
       "Content-Type": "multipart/form-data",
     },
   });
-  console.log(data);
   dispatch({ type: CREATE_PRODUCT_SUCCESS, payload: data.products });
 };
 
@@ -53,11 +50,12 @@ export const updateProduct = (formData, products) => async (dispatch) => {
 
 export const removeProduct = (product, products) => async (dispatch) => {
   dispatch({ type: DELETE_PRODUCT_REQUEST, payload: product });
-  const { data } = await axios.delete("/api/products/" + product._id);
+  //const { data } = await axios.delete("/api/products/" + product._id);
+
+  await axios.delete("/api/products/" + product._id);
 
   const filteredProducts = products.filter((item) => {
-    console.log(product, item);
-    return item._id != product._id;
+    return item._id !== product._id;
   });
 
   dispatch({ type: DELETE_PRODUCT_SUCCESS, payload: filteredProducts });
